ProyectoTitanic/requirements.txt
pandas
scikit-learn
joblib
django
djangorestframework
django-cors-headers